// src/api/models/sharingAgreementReferenceResponseStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Conluz API
 * Conluz is an API-driven application designed for the efficient management of an energy community,enabling the administration of community members and their corresponding supply points and the retrieval of consumption, production data.
 * OpenAPI spec version: 1.0.0
 */

/**
 * Status of the agreement: DRAFT, PUBLISHED or SUPERSEDED
 */
export type SharingAgreementReferenceResponseStatus = typeof SharingAgreementReferenceResponseStatus[keyof typeof SharingAgreementReferenceResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SharingAgreementReferenceResponseStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  SUPERSEDED: 'SUPERSEDED',
} as const;